Combined Rhett and Ryan's mqtt/chat
