+++ web/src/components/layout/Router.tsx
@@ -1,10 +1,12 @@
 import { Routes, Route } from "react-router-dom";
 import HomePage from "../../pages/HomePage";
+import ChatComponent from "../ChatComponent";
 
 export const RouterProvider = () => {
   return (
     <Routes>
       <Route path="/" element={<HomePage />} />
+      <Route path="chat" element={<ChatComponent />} />
     </Routes>
   );
 };
