--- web/src/mosquitto.conf
@@ -1,8 +1,0 @@
-# MQTT listener on port 1883 (standard MQTT)
-listener 1883
-protocol mqtt
-
-# WebSocket listener on port 9001 (for React)
-listener 9001
-protocol websockets
-allow_anonymous true
